fix: re-apply all MUGEN compat fixes (orphaned explods, hitonce, palfx, AIR, ReversalDef)
Re-applies all fixes from previous sessions that were lost in env reset:

1. DestroySelf removeexplods=0 orphans explods (MUGEN-spec):
   - orphanExplodsForPlayer sets mIsOrphaned=1, mPlayer=nullptr
   - destroyPlayerWithFlags(p, flag) handles both cases
   - Cached mOwnerRootID/mOwnerShadowOffsetY for rendering

2. HitDef hitonce: default 0, throw attrs default 1 (MUGEN 1.1 spec)

3. HitDef palfx.color/sinadd/invertall: reuse existing PalFX runtime
   - New storage fields in PlayerHitData
   - New parser fields in HitDefinitionController
   - setPlayerHitPaletteEffects passes all 6 palfx params

4. AIR: atoi->atof for fractional angle values (MUGEN 1.1)

5. ReversalDef: mark damage/getpower/givepower as project extensions
   in code comments (NOT MUGEN 1.1 spec)

All files compile cleanly with emscripten 6.0.9. Full WASM build
succeeds (3.3MB game.wasm).

+++ engine/DolmexicaInfinite/addons/prism/mugenanimationreader.cpp
@@ -225,7 +225,7 @@ namespace prism {
 
 		e->mScaleX = vectorElement->mVector.mSize >= 8 ? atof(vectorElement->mVector.mElement[7]) : 1.0;
 		e->mScaleY = vectorElement->mVector.mSize >= 9 ? atof(vectorElement->mVector.mElement[8]) : 1.0;
-		e->mAngleRad = vectorElement->mVector.mSize >= 10 ? degreesToRadians(double(atoi(vectorElement->mVector.mElement[9]))) : 0.0;
+		e->mAngleRad = vectorElement->mVector.mSize >= 10 ? degreesToRadians(atof(vectorElement->mVector.mElement[9])) : 0.0;
 
 		e->mInterpolateOffset = 0;
 		e->mInterpolateBlend = 0;
